improve model ủe

+++ app/Http/Controllers/usersController.php
@@ -13,14 +13,14 @@ class usersController extends Controller
 	{
 		$user=User::all();
 		return view("admin.users.list",["user"=>$user]);
-	}  
+	}
 	public function getAdd()
 	{
 		$city=city::all();
 		return view("admin.users.add",["city"=>$city]);
-	}  
+	}
 	public function postAdd(Request $r)
-	{				
+	{
 		$user=new User();
 		$user->name=$r->name;
 		$user->email=$r->email;
@@ -29,7 +29,7 @@ class usersController extends Controller
 		$u=$r->check;
 		if($r->changeOption)
 		{
-			
+
 			$user->phone_number=$r->phone_number;
 			$user->id_city=$r->id_city;
 			$user->id_district=$r->id_district;
@@ -41,8 +41,8 @@ class usersController extends Controller
 		$user->save();
 		return redirect("admin/users/add")->with("thongbao","Thêm thành công");
 
-	} 
-	
+	}
+
 	public function getEdit($id)
 	{
 		$user=user::find($id);
@@ -57,12 +57,12 @@ class usersController extends Controller
 
 		if($r->changePass)
 		{
-			 
+
 			$user->password=bcrypt($r->pass);
 		}
 		if($r->changeOption)
 		{
-			
+
 			$user->phone_number=$r->phone_number;
 			$user->id_city=$r->id_city;
 			$user->id_district=$r->id_district;
@@ -77,9 +77,12 @@ class usersController extends Controller
 	public function getDelete($id)
 	{
 		$user=user::find($id);
+        if($user->bill()->first())
+            return redirect("admin/users/list")->with("loi","Xóa thất bại vì user này giao dịch trong các hoá đơn");
+        $user->social()->delete();
 		$user->delete();
 		return redirect("admin/users/list")->with("thongbao","Xóa thành công");
 	}
-	  
+
 }
